Updated based on .NET MAUI testing.
