+++ sample/DataTheorem.TrustKit.Maui/MainPage.xaml
@@ -1,8 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="TrustKit.Maui.MainPage">
-    <StackLayout>
+             xmlns:maui="clr-namespace:TrustKit.Maui"
+             x:Class="TrustKit.Maui.MainPage"
+             x:DataType="maui:MainPageModel"
+             BindingContext="{maui:MainPageModel}">
+    <StackLayout Spacing="12" Margin="12">
         <Button Text="Test ShouldAllowConnection" Command="{Binding GetContentCommand}" CommandParameter="0" />
         <Button Text="Test ShouldBlockConnection" Command="{Binding GetContentCommand}" CommandParameter="1" />
         <Button Text="Test DomainNotPinned" Command="{Binding GetContentCommand}" CommandParameter="2" />
